(There must be something in the air, I started fiddling with this a couple of days ago and now @vesim987 has a stub loader in #1609. :).)

## Background

I got a bit obsessed trying to find ways to improve the performance impact from the RP2040 flash fixes applied in #1449 (fixing #1364). 

I originally tried writing a stub loader to read/write DR0, however I couldn't make it particularly fast. I think this is because the SPI flash abstraction is low-level so it runs a lot of small operations (Write Enable, Read Status, etc). Running the stub loader requires pushing/popping the full CPU register state each time. Splitting out large and small SPI commands, so only large ones use the stub, would probably help but will increase code size... I'm interested to see what @vesim987 comes up with, though, maybe moving all of the SPI register operations into the stub will be faster than I imagined. :)

## Detailed description

This draft PR is an experiment in using the RP2040's DMA Controller to manage SPI flash operations, instead:

1. Write data to target RAM (similar to when using a stub).
2. Configure DMA controller to send them to the SPI flash and read any results.
3. Wait for DMA transfers to complete.
4. Read back any read data from target RAM.

I hoped that this would need fewer SWD operations, and both "large" and "small" commands could use the same code path.

It's still not as fast as I'd like it to be, though... :thinking: 

## Performance

| Commit | stlink target | hosted target (+ DAP-Link adapter) | 
| ---- | ---- | ---- |
| current main | 2KB/s | 1KB/s |
| pre-#1449 | 49KB/s | |
| this PR | 18KB/s | 10KB/s |
| this PR + `mon freq 60000000` | 26KB/s |

(Benchmarked by running `load` in gdb with a 300KB-ish Micropython rp2 ELF file, looking at the speed reported by gdb.)

## Code size impact

At the moment this patch adds 192 bytes to the `stlink` binary size. This at least keeps it smaller than pre-#1449 (I think about 500 bytes saved there). But some things are not yet implemented...

## TODOs

* [ ] Save and restore the DMA controller register state.
* [ ] Do we need to save and restore the working area RAM? I noticed the EFM32 and LMI target code doesn't do this, but wanted to ask.
* [ ] Clean up a bunch of commented out debugging code.
* [ ] For some weird reason, SFDP reads return all 00s with this branch. I can't work out why, other read commands seem to work. Have ordered an RP2040 board with a bigger flash chip (SOIC-8) so I can easily attach a logic analyzer clip and see exactly what is going on.
* [ ] The RP2040 datasheet implies that the hardware can manage the CS pin, which would remove two SWD operations if so. (Maybe not in bidirectional mode though, idk.)

## Your checklist for this pull request

* [x] I've read the [Code of Conduct](https://github.com/blackmagic-debug/blackmagic/blob/main/CODE_OF_CONDUCT.md)
* [x] I've read the [guidelines for contributing](https://github.com/blackmagic-debug/blackmagic/blob/main/CONTRIBUTING.md) to this repository
* [x] It builds for hardware native (`make PROBE_HOST=native`)
* [x] It builds as BMDA (`make PROBE_HOST=hosted`)
* [ ] I've tested it to the best of my ability
* [x] My commit messages provide a useful short description of what the commits do