@projectgus 2023-08-24
I had another thought, that maybe the better way to do this is to split the `cortexm_run_stub` function into `start stub` and `finish stub` functions, so the stub starts running on CPU when the flash session starts and doesn't stop until the flash session is being cleaned up.

Then it can spin on a "control" memory word (or register) that communicates "write N bytes from the buffer, then read N bytes into the buffer" each time. This potentially gets the per-transfer overhead down to just the length of the bytes transferred plus the control word... :thinking:

(The other thought I had is that perhaps at this point the performance bottleneck is doing individual sector erases, so implementing block erase might be a big boost.)