@dragonmux 2024-02-02
Given we hit 12KB/s from BMDA with vesim's PR now it's been straightened out and that's what's landed on `main`, we're going to respectfully close this PR - if you can show this approach can actually be faster using less Flash then you'd be welcome to open a new PR to replace the speed up approach we've just landed.