@dragonmux 2024-02-02
No worries, thank you for proposing a very interesting and novel approach to the problem and taking the time to share 🙂